requirements.txt
ansible==2.9.6
appdirs==1.4.3
application-utility==1.3.2
asn1crypto==1.3.0
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.4
Brlapi==0.7.0
btrfsutil==1.2.0
CacheControl==0.12.6
cached-property==1.5.1
caffeine-ng==3.4.2
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.1
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.8
distlib==0.3.0
distro==1.4.0
docker==4.2.0
docker-compose==1.25.5
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
ewmh==0.1.6
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.8
gufw==19.10.0
html5lib==1.0.1
idna==2.9
importlib-metadata==1.5.0
iotop==0.6
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isc==2.0
itsdangerous==1.1.0
Jade-Application-Kit==3.0.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterthemes==0.20.0
kiwisolver==1.2.0
lensfun==0.3.95
lesscpy==0.14.0
libvirt-python==5.8.0
louis==3.13.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.2.0
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
npyscreen==4.10.5
numpy==1.18.2
ordered-set==3.1.1
packaging==20.3
pacman-mirrors==4.16.4
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.6.2
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
ply==3.11
progress==1.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pwquality==1.4.2
pycairo==1.19.1
pycparser==2.20
pycups==1.9.74
pycurl==7.43.0.5
PyGithub==1.47
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
PyQtWebEngine==5.14.0
pyrsistent==0.16.0
pysmbc==1.0.20
python-dateutil==2.8.1
python-xlib==0.27
pytoml==0.1.21
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
rados==2.0.0
ranger-fm==1.9.3
rbd==2.0.0
reportlab==3.5.42
requests==2.23.0
retrying==1.3.3
rgw==2.0.0
Send2Trash==1.5.0
setproctitle==1.1.10
sip==4.19.22
six==1.14.0
soupsieve==1.9.5
team==1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
ufw==0.36
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
